feat(storage): add S3 storage implementation and integrate with citizen service
- Introduced S3 storage functionality using MinIO client.
- Updated citizen service to encrypt and upload agreements to S3.
- Modified user agreement model to include S3 key, bucket, encrypted DEK, and key version.
- Enhanced configuration to support S3 settings.
- Updated database schema and queries to accommodate new fields for user agreements.
- Refactored repository interfaces and implementations to handle new data types.
- Added NoOp storage for fallback when S3 is disabled.
- Adjusted error handling and logging for better traceability during S3 operations.

+++ platform/services/identity-service/internal/repository/citizen.go
@@ -68,7 +68,7 @@ func (r *citizenRepository) Create(ctx context.Context, citizen *model.Citizen) 
 		PeselHash:     citizen.PESELHash,
 		EncryptedData: citizen.EncryptedData,
 		EncryptedDek:  citizen.EncryptedDEK,
-		KeyVersion:    citizen.KeyVersion,
+		KeyVersion:    int32(citizen.KeyVersion),
 	})
 	if err != nil {
 		return fmt.Errorf("failed to insert citizen: %w", err)
@@ -187,7 +187,7 @@ func (r *citizenRepository) GetByID(ctx context.Context, userID uuid.UUID) (*mod
 		PESELHash:     row.PeselHash,
 		EncryptedData: row.EncryptedData,
 		EncryptedDEK:  row.EncryptedDek,
-		KeyVersion:    row.KeyVersion,
+		KeyVersion:    int(row.KeyVersion),
 		CreatedAt:     row.CreatedAt,
 	}, nil
 }
@@ -207,7 +207,7 @@ func (r *citizenRepository) GetByPESELHash(ctx context.Context, peselHash string
 		PESELHash:     row.PeselHash,
 		EncryptedData: row.EncryptedData,
 		EncryptedDEK:  row.EncryptedDek,
-		KeyVersion:    row.KeyVersion,
+		KeyVersion:    int(row.KeyVersion),
 		CreatedAt:     row.CreatedAt,
 	}, nil
 }
